python/requirements.txt
azure-mgmt-apimanagement==3.0.0
azure-identity==1.13.0
azure-search-documents==11.6.0b11
PyYAML==6.0.2
markitdown==0.1.1
python-dotenv==1.1.0
openapi-spec-validator==0.5.1
openai==1.77.0
jsonschema[format-nongpl]>=4.18.0